added check of ownerId when deleting a card

+++ controllers/cards.js
@@ -18,17 +18,18 @@ module.exports.createCard = (req, res) => {
 };
 
 module.exports.removeCard = (req, res) => {
-  Card.findByIdAndDelete(req.params.cardId)
-    .then((card) => {
-      if (card === null) {
-        res.status(404).send({ message: 'Запрашиваемый ресурс не найден' });
-        return;
-      }
-      res.send({ data: card });
+  Card.findById(req.params.cardId)
+    .then(async (card) => {
+      const userId = req.user._id;
+      const ownerId = card.owner._id.toString();
+      if (ownerId === userId) {
+        const element = await Card.findByIdAndDelete(req.params.cardId);
+        res.send({ data: element });
+      } else res.status(403).send({ message: 'Нет прав на удаление' });
     })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: `Не удалось удалить карточку с cardId - ${req.params.cardId}` });
+      if (err.name === 'CastError' || err.name === 'TypeError') {
+        res.status(400).send({ message: `Не удалось найти карточку с cardId - ${req.params.cardId}` });
       } else res.status(500).send({ message: 'На сервере произошла ошибка' });
     });
 };
